+++ flame-flask/api/__init__.py
@@ -1,0 +1,6 @@
+from flask import Flask
+from .User import user
+
+
+def register_routes(app:Flask): 
+    app.register_blueprint(user)
