+++ flame-flask/model/__init__.py
@@ -1,0 +1,7 @@
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy(
+    engine_options={"connect_args": {
+        "options": "-c timezone=utc"
+    }})
+
