+++ flame-flask/service/UserService.py
@@ -1,0 +1,8 @@
+from model.User import User
+
+
+class UserService:
+
+    @staticmethod
+    def create(user: User):
+        user.db_create()
